+++ src/main/kotlin/baseball/domain/Computer.kt
@@ -1,4 +1,4 @@
-package baseball.service
+package baseball.domain
 
 import camp.nextstep.edu.missionutils.Randoms
